rewrite MindMap loading of default hierarchy
add file_locked class to act as a boolean-returning
function and a context-manager to, respectively, check
the lock-status of a file, and mark a file as locked in
context-enter then mark as unlocked in context-exit.
Rewrite pymm.read to use file_locked to lock both file
and default-hierarchy-file when reading from anything.
Only load default-hiearchy in MindMap if not locked.