fix critical factory can_encode(): check class equality
Now we check exact class equality, rather than isinstance().
Since there is always at least a 1-to-1 correspondence
of factories to elements, can_encode should be able to
only accept an element if it exactly matches its
decoding_element variable. Previously it would check if
the element was an instanceof(), which meant that a newly
created DefaultFactory (whose decoding_element was
BaseElement) would override encoding of ALL other pre-defined
factories. Only auto-generated factories would be prefered to
it. Which meant that Node and Attribute were both being
encoded by that dumb factory. Changing the test to class
equality ensures that there is (in general) only a factory that
explicitly says it can decode a particular element will be used
to do so. This should work since factories for unclaimed
elements are auto-generated to inherit from a closest-matching
factory.