correct factory conversion trace test
encode and decode store function classes used in converting
a given tree of elements. This tests that the number of
pymm decoding_element's in the encoding and decoding
factories match. Obviously if you do delete something in the
tree, you will change the numbers. But for now, a straight
conversion results in a test-pass.