fix encode/decode binding method
wrap the convert-event using a function, rather than an inline
lambda. The inline-lambda was used inside a for-loop and relied
on closure to correctly call the convert function. Inside a
for-loop, closure variables only solidify once the for-loop ends,
thereby causing all the wrapped functions for any specific
element to call the same function for different encode/decode
events.
By wrapping with an outside function, the closure is not altered
and therefore the function wrapping works as desired.