rewrite can_decode fxn to accept only tag, attrib
Previously, can_decode was required to have knowledge of un-decoded
element (ElementTree.Element, in this case) in order to extract
attrib and tag information. By instead passing in attrib and tag
arguments directly, the can_decode function no longer needs to know
the type of element it is dealing with. This simplifies the fxn in
case another backend (instead of xml.etree) is used